set the initial_datapackage_prop
