+++ javascript/app-framework/src/hooks/index.ts
@@ -5,6 +5,7 @@
 export * from './useDataPackage.js'
 export * from './useDatasetMenuItems.js'
 export * from './useHelpOnMount.js'
+export * from './useLoadDataPackage.js'
 export * from './useOnOutsideClick.js'
 export * from './usePersistenceService.js'
 export * from './useTableBundle.js'
